[Brightbox] Remove dependency on .fog in test
This isolates the tests from the values in `Fog.credentials` which is
based on the contents of `~/.fog`

Underlying problem is that the code (and many others) rely on the Fog
global making it too tightly coupled to the environment hence having to
stub.

Full issue: https://github.com/fog/fog/issues/1390

This will probably be a repeating pattern we need to work on to DRY up
so every test from every provider doesn't include this setup workaround.