Encapsulate Config in class
`Fog.credentials` is basically a global Hash that holds everything.
Thanks to primitive obsession it means that the configuration is dumb
forcing decentralised code to each know little bits about it.

This hash is dereferenced when passed into a service so any values that
are shared or dynamic are duplicated or can fall out of sync. Examples
being credentials shared by several services of one provider or tokens
that can be revoked and refreshed globally.

`Fog::Brightbox::Config` is a centralised configuration that smartens up
the details and can be initialised itself by legacy credentials.