Add support for Brightbox Orbit Cloud Storage
This adds beta support for Orbit, Brightbox's cloud storage service
based on OpenStack Swift.

The majority of the code here is extracted from the existing
OpenStack storage provider rather than awaiting a stable common provider
that is being worked upon.

A number of tests will be merged upstream into `fog` since the shared
tests are in Shindo unlike this projects.