Use relative paths and reduce requires
When requiring `fog-brightbox` all modules will be available
right away but they will be loaded only when used so this will free
some memory as well. This might sound as something small but when
counted together with all other fog modules it becomes a big win.

Since we are moving to modular providers we need to rely on bundler
to load the files and constant for us. This means that we should stop
loading only the service we want (`require 'fog/brightbox/compute'`)
in favor of the entire gem (`require 'fog/brightbox'`). With this
approach we can use `autoload` and make sure that the user will load
only what he will use.
In the picture of one provider this might sound silly,
but as we are working with several provider gems this will be a life saver.
The user put `fog` in the Gemfile  and bundler will automatically loaded
only the main file of each provider. Once he starts using one service
or another the files will be loaded.
This behavior will be a real change for `fog` since we will drop
our current lazy loading mechanism in favor of bundler.