Default `list_accounts` to not include links
The number of links in the API response of the account listing can cause
performance issues for customers with several accounts and/or many
resources on those accounts.

For an index action this is very slow and inefficient so the default is
being changed here and server side to omit nested resources.

To restore the original behaviour for either `list_accounts` or the
`servers.all` methods pass a Hash of `{:nested => true}` as an argument.

This is then passed through to the API and triggers the old behaviour
server side.