Changes to `#create_temp_url` with missing URL
The `management_url` is normally read from the authentication response
but an error would occur if called out of sequence. However, the error
was unclear relating to `NilClass`.

The method at fault was buried under the "requests" area and was implied
it was used during HTTP requests and so authentication must always have
happened. This moves the source of the method to the main service file.

This now raises `Fog::Brightbox::Storage::ManagementUrlUnknown` if the
`management_url` has not been either configured OR set by
authentication.

The `:brightbox_storage_management_url` is also "recognized" to not give
a warning if set as part of a configuration.

GH-41