Remove deprecated behaviours
A number of requests were implemented without needing to specify
identifiers which implied "details for the current user/account".
However, it added ambiguity when multiple users and accounts were
introduced and the behaviour has been deprecated with clearer, explicit
alternatives.

With v1.0 being a major release, this removes the behaviour fully.