Correct service/provider namespacing
Before solid guidance was available, the namespacing of services and
providers was not clear as to where the divisions lay.

This unfortunately meant there were two ways to create a provider and we
picked the one that was ultimately deprecated.

Compatibily code was added to support both `Fog::(service)::Brightbox` and
`Fog::Brightbox::(service)` but the latter is the prefered form since it
isolates the providers code into their own namespace.

Since this is a major change, this is the time to correct the original
oversight.