Fix "deprecation" warnings due to autoloading late
The `Fog.service` declaration has changed in later versions of
`fog-core` and attempts to `eval` constants to see if they exist and
fall back to the older naming scheme if not.

However, it used to work as in our case, the service was declared and
autoloading set up afterwards.

This changes the order to avoid issues with the core behaviour change.