Remove explicit bundler installation
This was a workaround for earlier versions of Ruby where bundler was not
included but since a number of them are failing because it is the
incorrect version, we may as well remove it.

Then we can set up a build matrix as required to keep the older versions
happy.