Use `example.test` for testing domains
Whilst `example.com` is a reserved domain per RFC2606 it was updated in
RFC6761 (https://tools.ietf.org/html/rfc6761) to be intended for
documentation rather than testing.

The `example.test` domain is intended for testing and should not be
routable by DNS so using the domain is preferred to avoid unintended
traffic.