Add `status` check to `ImageSelector`
The image selector, when given a list of images filters them to find the
most suitable type to use in a test.

The images API returns deleted resources for 1 hour so they can be seen
(and that their state has changed) before fading away.

A rare failure happened internally when the `official_minimal` image was
decommisioned within an hour of the test running and so the helper chose
an image to test with that immediately returned an error when trying to
create a server from the removed image.

This adds a check for `status == "available"` in the filtering so that
the helper will get an image that will not immediately error.