Fix creation of database servers from snapshot
The `snapshot_id` was not set up on the model which meant that passing
in a snapshot ID was incorrectly creating a new instance rather than
recreating a database from a snapshot.

Without the attribute, it was filtering out the required option in the
real API call but that calls would work if performed using requests.