Support two factor authention 2FA opt-in
This adds helper methods to opt-in connections to enable Two Factor
Authentcation (2FA) as part of the `Compute` service.

When creating a new Compute service, code can opt-in to receiving:

    Fog::Brightbox::OAuth2::TwoFactorMissingError

...instead of the existing `Excon::Error::Unauthorized` when the user
has 2FA enabled on their account and did not include the OTP code during
authentication.

The opt-in is done with `brightbox_support_two_factor: true`. The OTP
can be passed in using `brightbox_one_time_password`.

The opt-in is so a new error is not raised in client code unexpectedly.

Example (without opt-in):

    @options = {
      brightbox_client_id: "app-12345",
      brightbox_secret: "1234567890",
      brightbox_username: "jason.null@brightbox.com",
      brightbox_password: "HR4life"
    }
    @service = Fog::Brightbox::Compute.new(@options)
    @service.get_access_token! # => Excon::Error::Unauthorized

Example (with opt-in) without OTP:

    @options = {
      brightbox_client_id: "app-12345",
      brightbox_secret: "1234567890",
      brightbox_username: "jason.null@brightbox.com",
      brightbox_password: "HR4life",
      brightbox_support_two_factor: true
    }
    @service = Fog::Brightbox::Compute.new(@options)
    @service.get_access_token! # => Fog::Brightbox::OAuth2::TwoFactorMissingError

Example (with opt-in) with OTP:

    @options = {
      brightbox_client_id: "app-12345",
      brightbox_secret: "1234567890",
      brightbox_username: "jason.null@brightbox.com",
      brightbox_password: "HR4life",
      brightbox_support_two_factor: true,
      brightbox_one_time_password: @expected_otp
    }
    @service = Fog::Brightbox::Compute.new(@options)
    @service.get_access_token! # => Successful auth