Implement `..Storage::Directory#public_url`
This was not implemented in the OpenStack code we used as the template
for the storage server (OpenStack compatible) when we set up the code.

As such, attempting to generate a `File#public_url` would fail as it
relied on the directory URL being generated as well. The generated
string was only the object name.

Now it is correctly, the full URL (String) to access an object (if
public permissions are set).