Allow custom volume size in server creation
There is now an extended volumes API which allows passing a custom
volume size to network based storage. Along with attaching existing
volumes, that gives three variants that need to be passed through when
called with specific arguments.

Argument logic is not enforced here but left up to the API so mututally
exclusive arguments may be sent.