Fix `Fog::Brightbox::Directory#save` permissions
An error in the format of keys types (strings verses symbols) resulted
in the headers to set ACLs on a storage container to be set omitted
getting set as `nil`.

This results in any ACLs set using the library not being set on the
model and updating using `#save` being ignored.