Update DBS test for new API behaviour
Previously the API would only return the admin password during creation
as it was transient. This has been revised and is now stored encrypted
and is returned in other API requests such as a GET request.

The test is therefore incorrectly reporting errors against the new
behaviour and needed updating.